tf_fast_bw_fsa_staircase small fix

+++ TFNativeOp.py
@@ -807,7 +807,7 @@ def tf_fast_bw_fsa_staircase(seq_lens, **opts):
 
   def tf_fast_bw_fsa_staircase_wrapper(seq_lens):
     fsa = fast_bw_fsa_staircase(seq_lens, **opts)
-    return fsa.edges, fsa.weights, fsa.start_end_states
+    return fsa.edges.astype("int32"), fsa.weights.astype("float32"), fsa.start_end_states.astype("int32")
 
   edges, weights, start_end_states = tf.py_func(
     tf_fast_bw_fsa_staircase_wrapper,
